\name{version}
\alias{version}
\title{Get the version of libNVVM}
\description{
  This function returns the major and minor version of the
  libNVVM library.
}
\usage{
version()
}
\value{
  A list with two elements named major and minor. Each element
  is an integer.
}
\author{
Duncan Temple Lang
}

\examples{
version()
}
\keyword{programming}

